+++ design-system/page-layout/src/common/types.ts
@@ -1,4 +1,9 @@
-import { ReactNode } from 'react';
+import {
+  ButtonHTMLAttributes,
+  ElementType,
+  ReactElement,
+  ReactNode,
+} from 'react';
 
 interface SlotProps {
   isFixed?: boolean;
@@ -11,10 +16,35 @@ export interface SlotHeightProps extends SlotProps {
   height?: number;
 }
 
+export type ResizeButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & {
+  isLeftSidebarCollapsed: boolean;
+  label: string;
+  testId?: string;
+};
+
 export interface SlotWidthProps extends SlotProps {
   shouldPersistWidth?: boolean;
   width?: number;
 }
+
+export type ResizeControlProps = {
+  testId?: string;
+  overrides?: {
+    ResizeButton?: {
+      render?: (
+        Component: ElementType<ResizeButtonProps>,
+        props: ResizeButtonProps,
+      ) => ReactElement;
+    };
+  };
+  resizeButtonLabel?: string;
+  onCollapse?: () => void;
+  onExpand?: () => void;
+  onResizeStart?: () => void;
+  onResizeEnd?: () => void;
+};
+
+export interface LeftSidebarProps extends SlotWidthProps, ResizeControlProps {}
 
 export type DimensionNames =
   | 'leftPanelWidth'