+++ design-system/page-layout/src/components/index.ts
@@ -4,6 +4,7 @@ export { default as RightPanel } from './slots/right-panel';
 export { default as LeftPanel } from './slots/left-panel';
 export { default as RightSidebar } from './slots/right-sidebar';
 export { default as LeftSidebar } from './slots/left-sidebar';
+export { default as LeftSidebarWithoutResize } from './slots/left-sidebar-without-resize';
 export { default as Banner } from './slots/banner';
 export { default as TopNavigation } from './slots/top-navigation';
 export { default as Content } from './slots/content';